tests: Move SocketEvent into ruffle_socket_format crate
