+++ tests/socket-format/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ruffle_socket_mock"
+name = "ruffle_socket_format"
 authors.workspace = true
 edition.workspace = true
 homepage.workspace = true
@@ -8,6 +8,5 @@ repository.workspace = true
 version.workspace = true
 
 [dependencies]
-tracing = { workspace = true}
-tracing-subscriber = { workspace = true }
-ruffle_socket_format = { path = "../socket-format" }
+serde = { version = "1.0.152", features = ["derive"] }
+serde_json = "1.0.91"
